+++ mod.ts
@@ -96,8 +96,7 @@ export * from "./mock/mock.ts";
 export * from "./mock/spy_on.ts";
 export * from "./mock/global_this.ts";
 
-export { defineExpect, expect } from "./expect/mod.ts";
-export type { Expected, MatcherMap } from "./expect/mod.ts";
+export * from "./expect/mod.ts";
 export { anything } from "./dummy/anything.ts";
 export { any } from "./dummy/any.ts";
 export { arrayContaining } from "./dummy/array_containing.ts";